refactor: derived store to read sns loaded data
