+++ frontend/src/lib/services/sns.services.ts
@@ -12,16 +12,15 @@ import {
 import { AppPath } from "../constants/routes.constants";
 import {
   snsProposalsStore,
-  snsSummariesStore,
+  snsQueryStore,
   snsSwapCommitmentsStore,
-} from "../stores/projects.store";
+} from "../stores/sns.store";
 import { toastsStore } from "../stores/toasts.store";
 import type { Account } from "../types/account";
 import type { SnsSwapCommitment } from "../types/sns";
 import type { QuerySnsSummary, QuerySnsSwapState } from "../types/sns.query";
 import { getLastPathDetail, isRoutePath } from "../utils/app-path.utils";
 import { toToastError } from "../utils/error.utils";
-import { concatSnsSummaries } from "../utils/sns.utils";
 import { getAccountIdentity } from "./accounts.services";
 import { loadProposalsByTopic } from "./proposals.services";
 import {
@@ -35,7 +34,7 @@ export const loadSnsSummaries = ({
 }: {
   onError: () => void;
 }): Promise<void> => {
-  snsSummariesStore.setLoadingState();
+  snsQueryStore.setLoadingState();
 
   return queryAndUpdate<[QuerySnsSummary[], QuerySnsSwapState[]], unknown>({
     request: ({ certified, identity }) =>
@@ -44,10 +43,7 @@ export const loadSnsSummaries = ({
         querySnsSwapStates({ certified, identity }),
       ]),
     onLoad: ({ response, certified }) =>
-      snsSummariesStore.setSummaries({
-        summaries: concatSnsSummaries(response),
-        certified,
-      }),
+      snsQueryStore.setResponse({ response, certified }),
     onError: ({ error: err, certified }) => {
       console.error(err);
 
@@ -56,7 +52,7 @@ export const loadSnsSummaries = ({
       }
 
       // hide unproven data
-      snsSummariesStore.setLoadingState();
+      snsQueryStore.setLoadingState();
 
       toastsStore.error(
         toToastError({
