+++ frontend/src/tests/lib/components/launchpad/Projects.spec.ts
@@ -2,6 +2,7 @@
  * @jest-environment jsdom
  */
 
+import { SnsSwapLifecycle } from "@dfinity/sns";
 import { render, waitFor } from "@testing-library/svelte";
 import Projects from "../../../../lib/components/launchpad/Projects.svelte";
 import {
@@ -10,14 +11,15 @@ import {
 } from "../../../../lib/services/sns.services";
 import {
   snsesCountStore,
-  snsSummariesStore,
+  snsQueryStore,
   snsSwapCommitmentsStore,
-} from "../../../../lib/stores/projects.store";
+} from "../../../../lib/stores/sns.store";
 import en from "../../../mocks/i18n.mock";
 import {
   mockSnsSummaryList,
   mockSnsSwapCommitment,
 } from "../../../mocks/sns-projects.mock";
+import { snsResponsesForLifecycle } from "../../../mocks/sns-response.mock";
 
 jest.mock("../../../../lib/services/sns.services", () => {
   return {
@@ -28,7 +30,7 @@ jest.mock("../../../../lib/services/sns.services", () => {
 
 describe("Projects", () => {
   beforeEach(() => {
-    snsSummariesStore.reset();
+    snsQueryStore.reset();
     snsSwapCommitmentsStore.reset();
   });
 
@@ -49,10 +51,9 @@ describe("Projects", () => {
   it("should not load data when already loaded", () => {
     const principal = mockSnsSummaryList[0].rootCanisterId;
 
-    snsSummariesStore.setSummaries({
-      summaries: mockSnsSummaryList,
-      certified: false,
-    });
+    snsQueryStore.setResponse(
+      snsResponsesForLifecycle({ lifecycles: [SnsSwapLifecycle.Open] })
+    );
     snsSwapCommitmentsStore.setSwapCommitment({
       swapCommitment: mockSnsSwapCommitment(principal),
       certified: false,
@@ -65,7 +66,7 @@ describe("Projects", () => {
   });
 
   it("should not load data if store-state is null (loading)", () => {
-    snsSummariesStore.setLoadingState();
+    snsQueryStore.setLoadingState();
     snsSwapCommitmentsStore.setLoadingState();
 
     render(Projects);
@@ -77,10 +78,18 @@ describe("Projects", () => {
   it("should render projects", () => {
     const principal = mockSnsSummaryList[0].rootCanisterId;
 
-    snsSummariesStore.setSummaries({
-      summaries: mockSnsSummaryList,
-      certified: false,
-    });
+    const lifecycles = [
+      SnsSwapLifecycle.Open,
+      SnsSwapLifecycle.Open,
+      SnsSwapLifecycle.Open,
+      SnsSwapLifecycle.Open,
+    ];
+
+    snsQueryStore.setResponse(
+      snsResponsesForLifecycle({
+        lifecycles,
+      })
+    );
     snsSwapCommitmentsStore.setSwapCommitment({
       swapCommitment: mockSnsSwapCommitment(principal),
       certified: false,
@@ -88,14 +97,14 @@ describe("Projects", () => {
 
     const { getAllByTestId } = render(Projects);
 
-    expect(getAllByTestId("card").length).toBe(mockSnsSummaryList.length);
+    expect(getAllByTestId("card").length).toBe(lifecycles.length);
   });
 
   it("should render a message when no projects available", () => {
     const principal = mockSnsSummaryList[0].rootCanisterId;
 
-    snsSummariesStore.setSummaries({
-      summaries: [],
+    snsQueryStore.setResponse({
+      response: [[], []],
       certified: false,
     });
     snsSwapCommitmentsStore.setSwapCommitment({
