+++ frontend/src/tests/lib/components/launchpad/Proposals.spec.ts
@@ -6,7 +6,7 @@ import { ProposalStatus, type ProposalInfo } from "@dfinity/nns";
 import { render, waitFor } from "@testing-library/svelte";
 import Proposals from "../../../../lib/components/launchpad/Proposals.svelte";
 import { listSnsProposals } from "../../../../lib/services/sns.services";
-import { snsProposalsStore } from "../../../../lib/stores/projects.store";
+import { snsProposalsStore } from "../../../../lib/stores/sns.store";
 import en from "../../../mocks/i18n.mock";
 import { mockProposalInfo } from "../../../mocks/proposal.mock";
