+++ frontend/src/tests/lib/utils/sns.utils.spec.ts
@@ -1,5 +1,5 @@
 import {
-  concatSnsSummaries,
+  mapAndSortSnsQueryToSummaries,
   concatSnsSummary,
 } from "../../../lib/utils/sns.utils";
 import {
@@ -13,19 +13,19 @@ import {
 describe("sns-utils", () => {
   describe("concat sns summaries", () => {
     it("should return empty for undefined summary", () => {
-      const summaries = concatSnsSummaries([[], []]);
+      const summaries = mapAndSortSnsQueryToSummaries([[], []]);
 
       expect(summaries.length).toEqual(0);
     });
 
     it("should return empty for undefined swap query", () => {
-      const summaries = concatSnsSummaries([[mockSummary], []]);
+      const summaries = mapAndSortSnsQueryToSummaries([[mockSummary], []]);
 
       expect(summaries.length).toEqual(0);
     });
 
     it("should return empty for undefined swap init", () => {
-      const summaries = concatSnsSummaries([
+      const summaries = mapAndSortSnsQueryToSummaries([
         [mockSummary],
         [
           {
@@ -44,7 +44,7 @@ describe("sns-utils", () => {
     });
 
     it("should return empty for undefined swap state", () => {
-      const summaries = concatSnsSummaries([
+      const summaries = mapAndSortSnsQueryToSummaries([
         [mockSummary],
         [
           {
@@ -63,7 +63,7 @@ describe("sns-utils", () => {
     });
 
     it("should return empty if no root id are matching between summaries and swaps", () => {
-      const summaries = concatSnsSummaries([
+      const summaries = mapAndSortSnsQueryToSummaries([
         [mockSummary],
         [
           {
@@ -82,7 +82,7 @@ describe("sns-utils", () => {
     });
 
     it("should concat summaries and swaps", () => {
-      const summaries = concatSnsSummaries([
+      const summaries = mapAndSortSnsQueryToSummaries([
         [mockSummary],
         [
           {
@@ -103,7 +103,7 @@ describe("sns-utils", () => {
 
   describe("sort sns summaries", () => {
     it("should sort summaries and swaps", () => {
-      const summaries = concatSnsSummaries([
+      const summaries = mapAndSortSnsQueryToSummaries([
         [mockSummary, mockSnsSummaryList[1]],
         [
           {
